+++ src/barrels/Types.barrel.ts
@@ -1,4 +1,3 @@
 export type { SnowflakeOptions } from '../types/SnowflakeOptions.type';
 export type { SnowflakeResolve } from '../types/SnowflakeResolve.type';
-export type { SymbolicOptions } from '../types/SymbolicOptions.type';
-export type { SymbolicResolve } from '../types/SymbolicResolve.type';
+export type { IncrementOptions } from '../types/IncrementOptions.type';