+++ src/defaults/SymbolicOptions.default.ts
@@ -1,7 +1,0 @@
-import type { SymbolicOptions } from '../types/SymbolicOptions.type';
-
-export const SymbolicOptionsDefault: SymbolicOptions = {
-  epoch: '2025-01-01T00:00:00.000Z',
-  place_id: 0,
-  charset: '0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz'
-};