+++ src/main.ts
@@ -1,4 +1,4 @@
 export { Snowflake } from './Snowflake.class';
-export { Symbolic } from './Symbolic.class';
+export { Increment } from './Increment.class';
 
 export type * from './barrels/Types.barrel';