+++ src/types/SnowflakeOptions.type.ts
@@ -1,4 +1,6 @@
 export type SnowflakeOptions = {
+  format?: 'numeric' | 'symbolic';
+  charset?: string;
   epoch?: string;
   place_id?: number;
 };