+++ src/types/SymbolicOptions.type.ts
@@ -1,5 +1,0 @@
-export type SymbolicOptions = {
-  epoch?: string;
-  place_id?: number;
-  charset?: string;
-};